requirements.txt
gunicorn
Flask
kafka-python
smtplib
